+++ src/BarcodeKit.h
@@ -24,5 +24,6 @@
 #include "BarcodeKit/Common.h"
 
 #include "BarcodeKit/Code128.h"
+#include "BarcodeKit/EAN.h"
 
 #include "barcodekit_version.h"